requirements.txt
faiss-cpu
langchain
numpy
pymupdf
python-dotenv
requests
sentence-transformers
streamlit
tqdm
scikit-learn
transformers
torch
peft
torchvision
pillow
docling
langchain-docling
bm25s[full]
lingua-language-detector
PyStemmer
jax[cuda12]
FlagEmbedding
accelerate
stopwordsiso
huggingface_hub[cli]